Extract ITextInputMethodImpl into a BrowserTopLevelImpl
