+++ api/Avalonia.Browser.nupkg.xml
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<!-- https://learn.microsoft.com/en-us/dotnet/fundamentals/package-validation/diagnostic-ids -->
+<Suppressions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <Suppression>
+    <DiagnosticId>CP0002</DiagnosticId>
+    <Target>M:Avalonia.Browser.AvaloniaView.get_IsComposing</Target>
+    <Left>baseline/net7.0/Avalonia.Browser.dll</Left>
+    <Right>target/net8.0-browser1.0/Avalonia.Browser.dll</Right>
+  </Suppression>
+  <Suppression>
+    <DiagnosticId>CP0008</DiagnosticId>
+    <Target>T:Avalonia.Browser.AvaloniaView</Target>
+    <Left>baseline/net7.0/Avalonia.Browser.dll</Left>
+    <Right>target/net8.0-browser1.0/Avalonia.Browser.dll</Right>
+  </Suppression>
+</Suppressions>
